Create ShopFactory file and export class
